+++ Changelog.md
@@ -10,6 +10,7 @@
 
 ### Bugfixes
 
+- Add host config to the docker run config - [#988](https://github.com/PrefectHQ/ui/pull/988)
 - Add secret name and value routes to limit and offset exceptions - [#955](https://github.com/PrefectHQ/ui/pull/955)
 
 ## 2021-08-05